feat: make class Node<T>File
