+++ Calculator/Calculator/ViewController.swift
@@ -14,13 +14,5 @@ class ViewController: UIViewController {
     }
 }
 
-class Node<T> {
-    var data: T?
-    var next: Node?
-    
-    init (data: T?, next: Node? = nil) {
-        self.data = data
-        self.next = next
-    }
-}
+
